+++ app/Http/Controllers/ModuleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Module;
+use App\Models\Course;
 use Illuminate\Http\Request;
 
 class ModuleController extends Controller
@@ -15,8 +16,9 @@ class ModuleController extends Controller
     public function index()
     {
         //
-        $course=Course::all();
-        return view('module.index'compact('course'));
+        $this->authorize('viewAny' Module::class);
+        $modules=Module::all();
+        return view('module.index',compact('modules'));
     }
 
     /**
@@ -27,8 +29,10 @@ class ModuleController extends Controller
     public function create()
     {
         //
-        $course=Course::all();
-        return view('module.create',compact('course'));
+        $this->authorize('create' Module::class);
+        $courses=Course::all();
+        $modules=Module::all();
+        return view('module.create',compact('courses','modules'));
     }
 
     /**
@@ -40,6 +44,7 @@ class ModuleController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create' Module::class);
         $input=$request->all();
         $module=Module::create($input);
         return redirect('/module/'.$module->id);
@@ -54,6 +59,7 @@ class ModuleController extends Controller
     public function show(Module $module)
     {
         //
+        $this->authorize('view' Module::class);
         $course=Course::all();
         return view('module.show', compact('course'));
     }
@@ -67,7 +73,8 @@ class ModuleController extends Controller
     public function edit(Module $module)
     {
         //
-        $course=Course::all();
+        $this->authorize('update' Module::class);
+        $module=Module::all();
         return view('module.edit', compact('course'));
     }
 
@@ -81,6 +88,7 @@ class ModuleController extends Controller
     public function update(Request $request, Module $module)
     {
         //
+        $this->authorize('update' Module::class);
         $input=$request->all();
         $module->update($input);
         return redirect('module/'.$module->id);
@@ -95,6 +103,7 @@ class ModuleController extends Controller
     public function destroy(Module $module)
     {
         //
+        $this->authorize('delete' Module::class);
         $module->delete();
         return redirect('/module');
     }
