+++ app/Http/Controllers/ResourceController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Resource;
+use App\Models\Page;
 use Illuminate\Http\Request;
 
 class ResourceController extends Controller
@@ -15,9 +16,10 @@ class ResourceController extends Controller
     public function index()
     {
         //
-        $page=Page::all();
-        $resource=Resource::all();
-        return view('resource.index',compact('page','resource'));
+        $this->authorize('viewAny' Resource::class);
+        $pages=Page::all();
+        $resources = Resource::all();
+        return view('resource.index',compact('pages','resources'));
     }
 
     /**
@@ -28,6 +30,7 @@ class ResourceController extends Controller
     public function create()
     {
         //
+        $this->authorize('create' Resource::class);
         $page=Page::all();
         $resource=Resource::all();
         return view('resource.create',compact('page','resource'));
@@ -42,6 +45,7 @@ class ResourceController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create' Resource::class);
         $input=$request->all();
         $resource=Resource::create($input);
         return redirect('/resource/'.$resource->id);
@@ -56,6 +60,7 @@ class ResourceController extends Controller
     public function show(Resource $resource)
     {
         //
+        $this->authorize('view' Resource::class);
         $page=Page::all();
         $resource=Resource::all();
         return view('resource.show',compact('page','resource'));
@@ -70,6 +75,7 @@ class ResourceController extends Controller
     public function edit(Resource $resource)
     {
         //
+        $this->authorize('update' Resource::class);
         $page=Page::all();
         return view('resource.edit',compact('page','resource'));
     }
@@ -84,6 +90,7 @@ class ResourceController extends Controller
     public function update(Request $request, Resource $resource)
     {
         //
+        $this->authorize('update' Resource::class);
         $input=$request->all();
         $resource->update('$input');
         return redirect('resource/'.$resource->id);
@@ -98,6 +105,7 @@ class ResourceController extends Controller
     public function destroy(Resource $resource)
     {
         //
+        $this->authorize('delete' Resource::class);
         $resource->delete();
         return redirect('/resource');
     }
