+++ app/Http/Controllers/StudentController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\student;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class StudentController extends Controller
@@ -15,8 +16,10 @@ class StudentController extends Controller
     public function index()
     {
         //
-        $user=User::all();
-        return view('student.index', compact('user'));
+         $this->authorize('viewAny', Student::class);
+        $users=User::all();
+        $students =Student::all();
+        return view('student.index', compact('users','students'));
     }
 
     /**
@@ -27,8 +30,11 @@ class StudentController extends Controller
     public function create()
     {
         //
-        $user=User::all();
-        return view('student.create',compact('user'));
+        $this->authorize('create', Student::class);
+       
+        $users=User::all();
+        $student =Student::all();
+        return view('student.create',compact('users','student'));
     }
 
     /**
@@ -40,6 +46,7 @@ class StudentController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Student::class);
         $input=$request->all();
         $student->create($input);
         return redirect('/student'.$student->id);
@@ -54,6 +61,7 @@ class StudentController extends Controller
     public function show(student $student)
     {
         //
+        $this->authorize('view', Student::class);
         $user=User::all();
         return view('student.show', compact('user'));
     }
@@ -67,6 +75,7 @@ class StudentController extends Controller
     public function edit(student $student)
     {
         //
+        $this->authorize('update', Student::class);
         $user=User::all();
         return view('student.edit', compact('user'));
     }
@@ -81,6 +90,7 @@ class StudentController extends Controller
     public function update(Request $request, student $student)
     {
         //
+        $this->authorize('update', Student::class);
         $input=$request->all();
         $student->update($input);
         return redirect('/student'.$student->id);
@@ -95,6 +105,7 @@ class StudentController extends Controller
     public function destroy(student $student)
     {
         //
+        $this->authorize('delete', Student::class);
         $student->delete();
         return redirect('/student');
     }
